requirements.txt
grpcio
grpcio-tools
grpcio-reflection
pandas
numpy
scikit-learn
joblib
